[aws-lib] extract S3ObjectsBy{Hash,Key} to objects
